add teacher/auth route
This is a simple example of how to create a route which requires authorization.
In order to create a route which needs authorization you need to pass a function to
route.VERB which will be function(req, res, next).
req and res are know to us, the next is a callback which should be called only
if the user is authorized. By calling next you call the route.VERB function