initial commit for tools.
The tools directory will contain stuff which doesn't belong in the node.js
application. The first one is a go application which will try to login to dionysos.teilar.gr
in order to get the student credentials.

This tool is written with go because after testing we needed something faster than node.js.
Libuv doesn't offer us good performance compared to go